requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.1.0
alembic==1.11.1
altair==4.1.0
anyio==3.7.0
apache-airflow==2.2.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
async-timeout==4.0.2
attrs==20.3.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachelib==0.10.2
cachetools==4.2.4
cattrs==1.5.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
croniter==1.0.15
cryptography==41.0.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
dnspython==2.3.0
docutils==0.16
efficientnet-pytorch==0.7.1
email-validator==1.3.1
entrypoints==0.3
exceptiongroup==1.1.2
fastapi==0.68.1
filelock==3.12.2
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
frozenlist==1.3.3
gitdb==4.0.9
GitPython==3.1.20
google-api-core==2.11.1
google-auth==2.22.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.53.0
grpcio-tools==1.53.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.3
imageio==2.9.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==2.0.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
lazy-object-proxy==1.9.0
lazy_loader==0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.3.4
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
nbclient==0.5.8
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.5
numpy==1.24.4
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
param==1.13.0
parso==0.8.2
pendulum==2.1.2
pickleshare==0.7.5
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
prison==0.2.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==4.21.12
psutil==5.9.5
pyarrow==6.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.7.4
pydeck==0.6.2
Pygments==2.15.1
PyJWT==1.7.1
Pympler==0.9
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyWavelets==1.1.1
pywin32==302
pywinpty==1.1.5
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
redis==4.6.0
requests==2.26.0
requests-toolbelt==1.0.0
rich==13.4.2
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.48
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stability-sdk==0.8.3
starlette==0.14.2
streamlit==1.24.1
streamlit-cropper==0.2.1
swagger-ui-bundle==0.0.9
sympy==1.12
tabulate==0.8.10
tenacity==8.2.2
termcolor==2.3.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
tifffile==2020.9.3
toml==0.10.2
toolz==0.11.2
torch==2.0.1
torchvision==0.15.2
tornado==6.1
traitlets==4.3.3
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.26.7
uvicorn==0.15.0
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
WTForms==2.3.3
yarl==1.9.2
zipp==3.15.0
